fix: rename proto and delete unused imports
