+++ Dockerfile
@@ -11,7 +11,7 @@ ENV NODE_ENV=production
 WORKDIR /usr/src/app
 COPY --from=0 /usr/src/app/package.json /usr/src/app/package-lock.json ./
 COPY --from=0 /usr/src/app/dist ./dist
-COPY --from=0 /usr/src/app/protos ./protos
+COPY --from=0 /usr/src/app/proto ./proto
 RUN npm install
 EXPOSE 8080
 CMD ["npm", "start"]
