+++ src/users/users.service.ts
@@ -5,7 +5,7 @@ import Spike from '../spike/spike.service';
 import { kartoffelURL } from '../config';
 import { KartoffelError, UserNotFoundError, ApplicationError, SpikeError } from '../utils/errors';
 
-export default class UsersService {
+export class Kartoffel {
     private axiosInstance: AxiosInstance;
     private SpikeService: Spike;
 
@@ -14,7 +14,7 @@ export default class UsersService {
         this.axiosInstance = axios.create();
         // If authentication is needed, adds an authorization header to the axios instance.
         if (process.env.SPIKE_REQUIRED === 'true') {
-            this.addAuthInterceptor(); // async function. but cannot await since its a constructor.
+            this.addAuthInterceptor();
         }
     }
 
@@ -110,7 +110,7 @@ export default class UsersService {
      * Adds an Authorization header with an updated authentication token
      * from spike to the axios instance to kartoffel.
      */
-    private async addAuthInterceptor(): Promise<void> {
+    private addAuthInterceptor() {
         this.axiosInstance.interceptors.request.use(async (config) => {
             try {
                 const token: string = await this.SpikeService.getToken();
